requirements.txt
Flask==2.0.1
Flask-And-Redis==1.0.0
requests==2.26.0